requirements.txt
streamlit
PyMuPDF
requests
Pillow
python-dotenv
tiktoken
nltk
python-docx
celery
redis
azure-storage-blob
